% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_totals.R
\name{add_totals_col}
\alias{add_totals_col}
\title{Append a totals column to a data.frame.}
\usage{
add_totals_col(dat, na.rm = TRUE)
}
\arguments{
\item{dat}{an input data.frame with at least one numeric column.}

\item{na.rm}{should missing values (including NaN) be omitted from the calculations?}
}
\value{
Returns a data.frame with a totals column containing row-wise sums.
}
\description{
This function excludes the first column of the input data.frame, assuming it's a descriptive variable not to be summed.  It also excludes other non-numeric columns.
}
\examples{
library(dplyr) # for the \%>\% pipe
mtcars \%>\%
  crosstab(am, cyl) \%>\%
  add_totals_col
}
