% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{add_totals_row}
\alias{add_totals_row}
\title{Append a totals row to a data.frame.}
\usage{
add_totals_row(dat, fill = "-", na.rm = TRUE)
}
\arguments{
\item{dat}{an input data.frame with at least one numeric column.}

\item{fill}{if there are more than one non-numeric columns, what string should fill the bottom row of those columns?}

\item{na.rm}{should missing values (including NaN) be omitted from the calculations?}
}
\value{
Returns a data.frame with a totals row, consisting of "Total" in the first column and column sums in the others.
}
\description{
This function is deprecated, use \code{adorn_totals} instead.
}
