% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_col_title.R
\name{adorn_col_title}
\alias{adorn_col_title}
\title{Add column name to the top of a two-way tabyl.}
\usage{
adorn_col_title(dat, placement = "top")
}
\arguments{
\item{dat}{a data.frame of class \code{tabyl}.}

\item{placement}{whether the column name should be added to the top of the tabyl in an otherwise-empty row \code{"top"} or appended to the already-present row name variable (\code{"combined"}).  The formatting in the \code{"top"} option has the look of base R's \code{table()}; it also wipes out the other column names, making it hard to further use the data.frame besides formatting it for reporting.  The \code{"combined"} option is more conservative in this regard.}
}
\value{
the input tabyl, augmented with the column title.
}
\description{
This function adds the column variable name to the top of a \code{tabyl} for a complete display of information.  This makes the tabyl prettier, but renders the data.frame less useful for further manipulation.
}
\examples{

mtcars \%>\%
  tabyl(am, cyl) \%>\%
  adorn_col_title(placement = "top")
}
