% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_count}
\alias{assert_count}
\title{Verify that a vector of values has the expected number of \code{TRUE} values}
\usage{
assert_count(x, n = 1)
}
\arguments{
\item{x}{A logical vector without \code{NA} values}

\item{n}{The expected number of \code{TRUE} values}
}
\value{
\code{x} if \code{sum(x) == n} or an informative error message otherwise
}
\description{
Verify that a vector of values has the expected number of \code{TRUE} values
}
\examples{
data.frame(A = 1:5) \%>\%
  dplyr::mutate(
    big_values = assert_count(A > 2, n = 3)
  )

my_data <- data.frame(name = c("Bill", "Sam"), birthdate = c("2024-05-22", "2024-05-22"))
my_data |>
  dplyr::mutate(
    birthdate =
      dplyr::case_when(
        assert_count(name == "Bill" & birthdate == "2024-05-22") ~ "2024-05-23",
        TRUE ~ birthdate
      )
  )
}
