% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_count_true.R
\name{assert_count_true}
\alias{assert_count_true}
\title{Verify that a vector of values has the expected number of \code{TRUE} values}
\usage{
assert_count_true(x, n = 1)
}
\arguments{
\item{x}{A logical vecotor without \code{NA} values}

\item{n}{The expected number of \code{TRUE} values}
}
\value{
\code{x} if \code{sum(x) == n} or an informative error message otherwise
}
\description{
Verify that a vector of values has the expected number of \code{TRUE} values
}
\examples{
data.frame(A = 1:5) \%>\%
  dplyr::mutate(
    big_values = assert_count_true(A > 2, n = 3)
  )
}
