% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bindable.R
\name{check_bindable}
\alias{check_bindable}
\title{Compare the column names and types of two data.frames.}
\usage{
check_bindable(df1, df2)
}
\arguments{
\item{df1}{the first data.frame to compare.}

\item{df2}{the second data.frame to compare.
#' @return Returns a list with information about differences in the two data.frames}
}
\description{
Do you have multiple data.frames that are supposed to be the same - but aren't?  Use this to see why your \code{dplyr::bind_rows()} call fails.
}
\examples{
library(dplyr)
check_bindable(mtcars \%>\%
                 mutate(cyl = as.factor(cyl),
                        new_var = "hi"),
               mtcars \%>\%
                 select(-mpg, -wt) \%>\%
                 rename(CARB = carb)
)
}
