% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dupes.R
\name{check_dupes}
\alias{check_dupes}
\title{Check if there are duplicates in the supplied columns}
\usage{
check_dupes(df, ..., level = "stop")
}
\arguments{
\item{df}{the input data frame}

\item{...}{unquoted variable names to search for duplicates}

\item{level}{What should happen if there are duplicates. Can be "stop", "warning", "message" or "logical". Logical returns TRUE or FALSE.}
}
\value{
If level == "logical", TRUE/FALSE, otherwise the data frame is returned silently for further piping
}
\description{
Check whether the supplied columns have duplicates and, 
if there are duplicates, handle at the supplied level.
}
\examples{
# Make sure no duplicate cars were added since the last time we loaded the data
mtcars \%>\%
 mutate(make = row.names(.)) \%>\%
 dups_check(make)

# Check whether any Species have the same average Sepal.Width
iris \%>\%
 group_by(Species) \%>\%
 summarise(ave_sepal_width = mean(Sepal.Width)) \%>\%
 dups_check(ave_sepal_width, level="warning") \%>\%
 select(Species) # Silly example of pipe-ability
}

