% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{chisq.test}
\alias{chisq.test}
\title{Apply stats::chisq.test to a two-way tabyl}
\usage{
chisq.test(dat, tabyl_results = TRUE, ...)
}
\arguments{
\item{dat}{a two-way tabyl}

\item{tabyl_results}{if TRUE, also return `observed`, `expected`, `residuals` and `stdres` as tabyl}

\item{...}{other parameters passed to stats::chisq.test}
}
\value{
The result is the same as the one of stats::chisqt.test. If `tabyl_results` 
is TRUE, the returned tables `observed`, `expected`, `residuals` and `stdres` 
are converted to tabyls.
}
\description{
This generic function overrides stats::chisq.test. If the passed table 
is a two-way tabyl, it runs it through janitor::chis.test.tabyl, otherwise
it just calls stats::chisq.test.
}
\examples{
tab <- tabyl(mtcars, gear, cyl)
chisq.test(tab)
chisq.test(tab)$residuals

}
