% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NA_variants.R
\name{clean_NA_variants}
\alias{clean_NA_variants}
\title{Convert common NA string values to true \code{NA} values, throughout a data.frame.}
\usage{
clean_NA_variants(dat, addl_strings = NULL)
}
\arguments{
\item{dat}{data.frame to operate on.}

\item{addl_strings}{an optional character vector of additional strings to convert.}
}
\value{
Returns a data.frame (or tbl_df if a tbl_df was supplied).
}
\description{
Converts strings like "#N/A", and any user-specified strings, into \code{NA}.

Default values to convert to NA are "NA", "#N/A", "N/A", "n/a", "#NAME?".  You can add your own additional values via a character vector; those values will also be converted to \code{NA}.
}
\examples{
clean_NA_variants(mtcars, "4") # a silly example;
# mtcars has no string NA values, but this will convert 4s to NA
}

