% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NA_variants.R
\name{clean_NA_vec}
\alias{clean_NA_vec}
\title{Turn common NA string values in a vector into true \code{NA} values.}
\usage{
clean_NA_vec(vec, na_vals)
}
\arguments{
\item{vec}{vector to operate on.}

\item{na_vals}{a character vector of strings to convert.}
}
\value{
Returns a vector.
}
\description{
Converts strings like "#N/A", and any user-specified strings, into \code{NA}.  Operates at vector level.

Supply a vector \code{na_vals} of strings to convert to \code{NA}.
}
\examples{
# not run:
# clean_NA_vec(letters, c("b", "d"))
}
\keyword{internal}

