% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df_types.R
\name{compare_df_types_class_detect}
\alias{compare_df_types_class_detect}
\alias{compare_df_types_class_detect.factor}
\alias{compare_df_types_class_detect.default}
\title{Describe the class(es) of an object}
\usage{
compare_df_types_class_detect(x)

\method{compare_df_types_class_detect}{factor}(x)

\method{compare_df_types_class_detect}{default}(x)
}
\arguments{
\item{x}{The object to describe}
}
\value{
A character scalar describing the class(es) of an object where if the
  scalar will match, columns in a data.frame (or similar object) should bind
  together without issue.
}
\description{
Describe the class(es) of an object
}
\details{
An S3 generic method can be written for
  \code{compare_df_types_class_detect()} for other types that may need more
  definition than the default method.
}
\section{Methods (by class)}{
\itemize{
\item \code{factor}: Describe factors with their levels
and if they are ordered.

\item \code{default}: List all classes of an object.
}}

\seealso{
Other Data frame type comparison: \code{\link{compare_df_types_success}},
  \code{\link{compare_df_types}}
}
\concept{Data frame type comparison}
