% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df_types.R
\name{compare_df_types_success}
\alias{compare_df_types_success}
\title{Will row binding succeed?}
\usage{
compare_df_types_success(..., return = "mismatch",
  bind_check = c("rbind", "bind_rows"), verbose = TRUE)
}
\arguments{
\item{...}{data.frames or similar objects.  The values may optionally be
named arguments; if named, the output column will be the name; if not
named, the output columne will be the data.frame name (see examples
section).}

\item{return}{Should a sumary of "all" columns be returned, only return
"match"ing columns, or only "mismatch"ing columns.}

\item{bind_check}{What method of binding shoudl be used to determine matches?
With "rbind", columns missing from a data.frame would be considered a
mismatch; with "bind_rows" columns missing from a data.frame would be
considered a match (as in \code{dplyr::bind_rows()}.}

\item{verbose}{Print the mismatching columns if binding will fail.}
}
\value{
\code{TRUE} if row binding will succeed or \code{FALSE} if it will
  fail.
}
\description{
Will row binding succeed?
}
\examples{
compare_df_types_success(data.frame(A=1), data.frame(A=2))
compare_df_types_success(data.frame(A=1), data.frame(B=2))
compare_df_types_success(data.frame(A=1), data.frame(B=2), verbose=FALSE)
compare_df_types_success(data.frame(A=1), data.frame(B=2), bind_check="bind_rows")
}
\seealso{
Other Data frame type comparison: \code{\link{compare_df_types_class_detect}},
  \code{\link{compare_df_types}}
}
\concept{Data frame type comparison}
