% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_dupes.R
\name{count_dupes}
\alias{count_dupes}
\title{Count the number of duplicates in the supplied columns}
\usage{
count_dupes(df, ..., sort = FALSE)
}
\arguments{
\item{df}{the input data frame}

\item{...}{unquoted variable names to search for duplicates}
}
\value{
the data frame is returned silently for further piping
}
\description{
Summarise the number of duplicates on the supplied columns.
}
\examples{
# Count duplicate ids
df <- data.frame(id = c(1,2,3,3,3), stuff = c("a", "b", "c", "d", "d"))
count_dupes(df)
count_dupes(df, id)

# Join the number of duplicates with the duplicate observations from get_dupes
df <- data.frame(id = c(1,2,3,3,3), stuff = c("a", "b", "c", "d", "d"))
t <- df \%>\%
  get_dupes(., id) \%>\%
  left_join(count_dupes(., id))
}

