% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dupes.R
\name{get_dupes}
\alias{get_dupes}
\title{Filter a df to all duplicated observations on the supplied columns}
\usage{
get_dupes(df, ...)
}
\arguments{
\item{df}{the input data frame}

\item{...}{unquoted variable names to search for duplicates}
}
\value{
A data frame filter to only duplicate observations
}
\description{
Filter a dataframe to all instances of on duplicates on
the supplied columns, such that you can \code{View()} the result
or take any other steps necessary to resolve the duplicates. 

This differs from \code{duplicated()} in that \code{duplicated()} 
returns one row per duplicate, while \code{get_dupes()} returns all
duplicates, and also all columns.
}
\examples{
dupe_df <- data.frame(a = c(1,3,3,3,5), b = c("a", "c","c", "e", "c"))
get_dupes(dupe_df)
get_dupes(dupe_df, a)
}

