% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fuzzy_dupes.R
\name{get_fuzzy_dupes}
\alias{get_fuzzy_dupes}
\title{Find clusters of records that almost match}
\usage{
get_fuzzy_dupes(dat, max_dist = 2)
}
\arguments{
\item{dat}{the data.frame to check for duplicates.  Right now takes only a one-vector data.frame.}

\item{max_dist}{maximum distance to count as a duplicate.  Increase for looser definition of duplicates (and potentially more false positives).}
}
\value{
a data.frame of duplicates, along with their cluster key (for easier cleaning up) and distance between the two instances.
}
\description{
Match near duplicates, based on text distance.  Quickly ID records that should possibly be identical, and then use the output to quickly clean up if needed.
}
\examples{
library(dplyr)
dat <- mtcars \%>\%
  transmute(cars = row.names(.))
get_fuzzy_dupes(dat, 1)
get_fuzzy_dupes(dat, 3)
}

