% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_to_one.R
\name{get_one_to_one}
\alias{get_one_to_one}
\title{Find the list of columns that have a 1:1 mapping to each other}
\usage{
get_one_to_one(dat)
}
\arguments{
\item{dat}{A data.frame or similar object}
}
\value{
A list with one element for each group of columns that map
  identically to each other.
}
\description{
Find the list of columns that have a 1:1 mapping to each other
}
\examples{
foo <- data.frame(
  Lab_Test_Long=c("Cholesterol, LDL", "Cholesterol, LDL", "Glucose"),
  Lab_Test_Short=c("CLDL", "CLDL", "GLUC"),
  LOINC=c(12345, 12345, 54321),
  Person=c("Sam", "Bill", "Sam"),
  stringsAsFactors=FALSE
)
get_one_to_one(foo)
}
