% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{mu_to_u}
\alias{mu_to_u}
\title{Constant to help map from mu to u}
\usage{
mu_to_u
}
\description{
This is a character vector with names of all known Unicode code points that
look like the Greek mu or the micro symbol and values of "u".  This is
intended to simplify mapping from mu or micro in Unicode to the character "u"
with \code{clean_names()} and \code{make_clean_names()}.
}
\details{
See the help in \code{clean_names()} for how to use this.
}
\seealso{
Other Set names:
\code{\link[=clean_names]{clean_names()}},
\code{\link[=find_header]{find_header()}},
\code{\link[=row_to_names]{row_to_names()}}
}
\concept{Set names}
