% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_skip_na.R
\name{paste_skip_na}
\alias{paste_skip_na}
\title{Concatenate strings dropping missing values}
\usage{
paste_skip_na(..., sep = " ", collapse = NULL)
}
\arguments{
\item{..., sep, collapse}{See \code{?paste}}
}
\value{
A character vector of pasted values.
}
\description{
Concatenate strings dropping missing values
}
\details{
If all values are missing, the value from the first argument is
  preserved.
}
\examples{
paste_skip_na(NA) # NA_character_
paste_skip_na("A", NA) # "A"
paste_skip_na("A", NA, c(NA, "B"), sep = ",") # c("A", "A,B")
}
