% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
Exported from the magrittr package.  To learn more, run \code{?magrittr::`\%>\%`}.
}
\examples{
mtcars \%>\%
  tabyl(carb, cyl) \%>\%
  adorn_totals()
}
\keyword{internal}
