% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Removes empty columns from a data.frame.}
\usage{
remove_empty_cols(dat)
}
\arguments{
\item{dat}{the input data.frame.}
}
\value{
Returns the data.frame with no empty columns.
}
\description{
This function is deprecated, use \code{\link[=remove_empty]{remove_empty("cols")}} instead.
}
\keyword{internal}
