% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_half_up.R
\name{round_half_up}
\alias{round_half_up}
\title{Round a numeric vector; halves will be rounded up, ala Microsoft Excel.}
\usage{
round_half_up(x, digits = 0)
}
\arguments{
\item{x}{a numeric vector to round.}

\item{digits}{how many digits should be displayed after the decimal point?}
}
\description{
In base R \code{round()}, halves are rounded to even, e.g., 12.5 and 11.5 are both rounded to 12.  This function rounds 12.5 to 13 (assuming \code{digits = 0}).  Negative halves are rounded away from zero, e.g., -0.5 is rounded to -1.

This may skew subsequent statistical analysis of the data, but may be desirable in certain contexts.  This function is implemented exactly from \url{https://stackoverflow.com/a/12688836}; see that question and comments for discussion of this issue.
}
\examples{
round_half_up(12.5)
round_half_up(1.125, 2)
round_half_up(1.125, 1)
round_half_up(-0.5, 0) # negatives get rounded away from zero

}
