% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_to_fraction.R
\name{round_to_fraction}
\alias{round_to_fraction}
\title{Round a value to a fraction, optionally then rounding to a number of digits.}
\usage{
round_to_fraction(x, denominator, digits = Inf)
}
\arguments{
\item{x}{A numeric vector}

\item{denominator}{The denominator of the fraction for rounding (a scalar or
vector positive integer).}

\item{digits}{Integer indicating the number of decimal places to be used
(passed \code{base::round()}). Negative values are allowed (see
Details). The number of digits for rounding after rounding to the
fraction (\code{Inf} indicates no subsequent rounding)}
}
\value{
x rounded to a decimal value that has an integer numerator relative
  to \code{denominator} (possibly subsequently rounded to a number of decimal
  digits).
}
\description{
Round a value to a fraction, optionally then rounding to a number of digits.
}
\details{
If \code{digits} is \code{Inf}, \code{x} is rounded to the fraction
  and then kept at full precision.  If \code{digits} is \code{"auto"}, the
  number of digits is automatically selected as
  \code{ceiling(log10(denominator)) + 1}.
}
\examples{
round_to_fraction(1.6, denominator=2)
round_to_fraction(pi, denominator=7) # 22/7
round_to_fraction(c(8.1, 9.2), denominator=c(7, 8))
round_to_fraction(c(8.1, 9.2), denominator=c(7, 8), digits=3)
round_to_fraction(c(8.1, 9.2, 10.3), denominator=c(7, 8, 1001), digits="auto")
}
