% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_and_un_tabyl.R
\name{un_tabyl}
\alias{un_tabyl}
\title{Remove \code{tabyl} attributes to a data.frame.}
\usage{
un_tabyl(dat)
}
\arguments{
\item{dat}{a data.frame of class \code{tabyl}.}
}
\value{
Returns the same data.frame, but without the \code{tabyl} class and attributes.
}
\description{
Strips away all \code{tabyl}-related attributes from a data.frame.
}
\examples{
mtcars \%>\%
  as_tabyl() \%>\%
  un_tabyl()
}
