% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_and_untabyl.R
\name{untabyl}
\alias{untabyl}
\title{Remove \code{tabyl} attributes from a data.frame.}
\usage{
untabyl(dat)
}
\arguments{
\item{dat}{a data.frame of class \code{tabyl}.}
}
\value{
Returns the same data.frame, but without the \code{tabyl} class and attributes.
}
\description{
Strips away all \code{tabyl}-related attributes from a data.frame.
}
\examples{

mtcars \%>\%
  tabyl(am) \%>\%
  untabyl() \%>\%
  attributes() # tabyl-specific attributes are gone
}
