% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_quotes.R
\name{zap_curly_quotes}
\alias{zap_curly_quotes}
\title{Zap curly quotes to standard.}
\usage{
zap_curly_quotes(x)
}
\arguments{
\item{x}{character string to process}
}
\value{
the processed string
}
\description{
Zap curly quotes to standard.
}
\examples{
x <- "it\\u2019s"
x
x == "it's"
zap_curly_quotes(x) == "it's"

}

