% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{adorn_crosstab}
\alias{adorn_crosstab}
\title{Add presentation formatting to a crosstabulation table.}
\usage{
adorn_crosstab(
  dat,
  denom = "row",
  show_n = TRUE,
  digits = 1,
  show_totals = FALSE,
  rounding = "half to even"
)
}
\arguments{
\item{dat}{a data.frame with row names in the first column and numeric values in all other columns.  Usually the piped-in result of a call to  \code{crosstab} that included the argument \code{percent = "none"}.}

\item{denom}{the denominator to use for calculating percentages.  One of "row", "col", or "all".}

\item{show_n}{should counts be displayed alongside the percentages?}

\item{digits}{how many digits should be displayed after the decimal point?}

\item{show_totals}{display a totals summary? Will be a row, column, or both depending on the value of \code{denom}.}

\item{rounding}{method to use for truncating percentages - either "half to even", the base R default method, or "half up", where 14.5 rounds up to 15.}
}
\value{
Returns a data.frame.
}
\description{
This function is deprecated, use the \code{adorn_} family of functions instead.
}
\keyword{internal}
