% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_ns.R
\name{adorn_ns}
\alias{adorn_ns}
\title{Add underlying Ns to a tabyl displaying percentages.}
\usage{
adorn_ns(dat, position = "rear")
}
\arguments{
\item{dat}{a data.frame of class \code{tabyl}, typically the result of a call to \code{adorn_percentages} on a \code{tabyl}.}

\item{position}{should the N go in the front, or in the rear, of the percentage?}
}
\value{
a data.frame with Ns appended
}
\description{
This function adds back the underlying Ns to a \code{tabyl} whose percentages were calculated using \code{adorn_percentages()}, to display the Ns and percentages together.
}
\examples{

mtcars \%>\%
  tabyl(am, cyl) \%>\%
  adorn_percentages("col") \%>\%
  adorn_pct_formatting() \%>\%
  adorn_ns(position = "front")
}
