% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn_rounding.R
\name{adorn_rounding}
\alias{adorn_rounding}
\title{Round the numeric columns in a data.frame.}
\usage{
adorn_rounding(dat, digits = 1, method = "half to even",
  skip_first_col = TRUE)
}
\arguments{
\item{dat}{a \code{tabyl} or other data.frame with a tabyl-like layout.}

\item{digits}{how many digits should be displayed after the decimal point?}

\item{method}{method to use for truncating percentages - either "half to even", the base R default method, or "half up", where 14.5 rounds up to 15.}

\item{skip_first_col}{should the first column be left un-rounded, assuming it contains values of a descriptive variable as in a \code{tabyl}?  Defaults to \code{TRUE}.}
}
\value{
Returns the data.frame with rounded numeric columns.
}
\description{
This function defaults to excluding the first column of the input data.frame, assuming that it contains a descriptive variable.  Designed to run on a \code{tabyl} data.frame, but can run on any data.frame with at least one numeric column.
}
