% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Cleans names of a data.frame.}
\usage{
clean_names(dat)
}
\arguments{
\item{dat}{the input data.frame.}
}
\value{
Returns the data.frame with clean names.
}
\description{
Resulting names are unique and consist only of the \code{_} character, lowercase letters, and numbers.

Accented characters are
transliterated to ASCII.  For example, an "o" with a german umlaut over it becomes "o", and the Spanish character "enye" becomes "n".
}
\examples{
# not run:
# clean_names(poorly_named_df)

# or with the pipe character from dplyr:
# poorly_named_df \%>\% clean_names()

# not run:
# library(readxl)
# readxl("messy_excel_file.xlsx") \%>\% clean_names()
}
