% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df_cols.R
\name{compare_df_cols_same}
\alias{compare_df_cols_same}
\title{Will row binding succeed?}
\usage{
compare_df_cols_same(..., return = "mismatch", bind_method = c("rbind",
  "bind_rows"), verbose = TRUE)
}
\arguments{
\item{...}{A combination of data.frames, tibbles, and lists of
data.frames/tibbles.  The values may optionally be named arguments; if
named, the output column will be the name; if not named, the output columne
will be the data.frame name (see examples section).}

\item{return}{Should a summary of "all" columns be returned, only return
"match"ing columns, or only "mismatch"ing columns?}

\item{bind_method}{What method of binding should be used to determine
matches? With "rbind", columns missing from a data.frame would be
considered a mismatch (as in \code{base::rbind()}; with "bind_rows",
columns missing from a data.frame would be considered a match (as in
\code{dplyr::bind_rows()}.}

\item{verbose}{Print the mismatching columns if binding will fail.}
}
\value{
\code{TRUE} if row binding will succeed or \code{FALSE} if it will
  fail.
}
\description{
Will row binding succeed?
}
\examples{
compare_df_cols_same(data.frame(A=1), data.frame(A=2))
compare_df_cols_same(data.frame(A=1), data.frame(B=2))
compare_df_cols_same(data.frame(A=1), data.frame(B=2), verbose=FALSE)
compare_df_cols_same(data.frame(A=1), data.frame(B=2), bind_method="bind_rows")
}
\seealso{
Other Data frame type comparison: \code{\link{compare_df_cols}},
  \code{\link{describe_class}}
}
\concept{Data frame type comparison}
