% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Generate a crosstabulation of two vectors.}
\usage{
crosstab(vec1, vec2, percent = "none", show_na = TRUE)
}
\arguments{
\item{vec1}{the vector to place on the crosstab column.}

\item{vec2}{the vector to place on the crosstab row.}

\item{percent}{which grouping to use for percentages, if desired (defaults to counts).}

\item{show_na}{should cases where both variables are NA be omitted?}
}
\value{
Returns a data.frame (actually a \code{tbl_df}) with the frequencies of the crosstabulatd variables.
}
\description{
Create a crosstab, displaying either frequencies or percentages calculated by row, column, or overall.  Vectors don't have to be from the same data.frame, but typically are.
}
\examples{
crosstab(mtcars$cyl, mtcars$gear)
crosstab(mtcars$cyl, mtcars$gear, "row")

}

