% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{crosstab}
\alias{crosstab}
\title{Generate a crosstabulation of two vectors.}
\usage{
crosstab(...)
}
\arguments{
\item{...}{arguments}
}
\description{
This function is deprecated, use \code{tabyl(dat, var1, var2)} instead.
}
\keyword{internal}
