% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{fisher.test}
\alias{fisher.test}
\alias{fisher.test.default}
\alias{fisher.test.tabyl}
\title{Apply \code{stats::fisher.test()} to a two-way tabyl}
\usage{
fisher.test(x, ...)

\method{fisher.test}{default}(x, y = NULL, ...)

\method{fisher.test}{tabyl}(x, ...)
}
\arguments{
\item{x}{A two-way tabyl, a numeric vector or a factor}

\item{...}{Parameters passed to \code{\link[stats:fisher.test]{stats::fisher.test()}}}

\item{y}{if x is a vector, must be another vector or factor of the same length}
}
\value{
The same as the one of \code{stats::fisher.test()}.
}
\description{
This generic function overrides \code{\link[stats:fisher.test]{stats::fisher.test()}}. If the passed table
is a two-way tabyl, it runs it through \code{janitor::fisher.test.tabyl}, otherwise
it just calls \code{stats::fisher.test()}.
}
\examples{
tab <- tabyl(mtcars, gear, cyl)
fisher.test(tab)

}
