% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor.R
\docType{package}
\name{janitor-package}
\alias{janitor}
\alias{janitor-package}
\title{janitor: Simple Tools for Examining and Cleaning Dirty Data}
\description{
The main janitor functions can: perfectly format data.frame column names and explore duplicate records. This package follows the principles of the "tidyverse" and works well with the pipe function %>%. janitor was built with beginning-to-intermediate R users in mind and is optimized for user-friendliness.
}
\section{Package context}{

Advanced users can do most things covered here, but they can do it
faster with janitor and save their thinking for more fun tasks.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sfirke/janitor}
  \item \url{https://sfirke.github.io/janitor/}
  \item Report bugs at \url{https://github.com/sfirke/janitor/issues}
}

}
\author{
\strong{Maintainer}: Sam Firke \email{samuel.firke@gmail.com}

Other contributors:
\itemize{
  \item Bill Denney \email{wdenney@humanpredictions.com} [contributor]
  \item Chris Haid \email{chrishaid@gmail.com} [contributor]
  \item Ryan Knight \email{ryangknight@gmail.com} [contributor]
  \item Malte Grosser \email{malte.grosser@gmail.com} [contributor]
  \item Jonathan Zadra \email{jonathan.zadra@sorensonimpact.com} [contributor]
  \item Olivier Roy [contributor]
  \item Josep Pueyo-Ros \email{josep.pueyo@udg.edu} [contributor]
}

}
\keyword{internal}
