% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor.R
\docType{package}
\name{janitor-package}
\alias{janitor}
\alias{janitor-package}
\title{janitor}
\description{
janitor has simple little tools for examining and cleaning dirty data.
}
\section{Main functions}{

The main janitor functions can: perfectly format data.frame
column names; provide quick counts of variable combinations (i.e.,
frequency tables and crosstabs); and explore duplicate records. Other
janitor functions nicely format the tabulation results. These
tabulate-and-report functions approximate popular features of SPSS and
Microsoft Excel.
}

\section{Package context}{

This package follows the principles of the "tidyverse" and works
well with the pipe function \verb{\%>\%}.

janitor was built with beginning-to-intermediate R users in mind
and is optimized for user-friendliness.  Advanced users can do most
things covered here, but they can do it faster with janitor and save
their thinking for more fun tasks.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sfirke/janitor}
  \item \url{https://sfirke.github.io/janitor/}
  \item Report bugs at \url{https://github.com/sfirke/janitor/issues}
}

}
\author{
\strong{Maintainer}: Sam Firke \email{samuel.firke@gmail.com}

Other contributors:
\itemize{
  \item Bill Denney \email{wdenney@humanpredictions.com} [contributor]
  \item Chris Haid \email{chrishaid@gmail.com} [contributor]
  \item Ryan Knight \email{ryangknight@gmail.com} [contributor]
  \item Malte Grosser \email{malte.grosser@gmail.com} [contributor]
  \item Jonathan Zadra \email{jonathan.zadra@sorensonimpact.com} [contributor]
  \item Olivier Roy [contributor]
}

}
