% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{janitor_deprecated}
\alias{janitor_deprecated}
\title{Deprecated Functions in Package janitor}
\description{
These functions are provided for compatibility with older versions of
\code{janitor} only, and may be defunct as soon as the next release.
}
\details{
\itemize{
  \item \code{\link{use_first_valid_of}}
  \item \code{\link{convert_to_NA}}
}
}

