% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{janitor_deprecated}
\alias{janitor_deprecated}
\title{Deprecated Functions in Package janitor}
\description{
These functions have already become defunct or may be defunct as soon as the next release.
}
\details{
\itemize{
  \item \code{\link{adorn_crosstab}}
  \item \code{\link{crosstab}}
  \item \code{\link{use_first_valid_of}}
  \item \code{\link{convert_to_NA}}
  \item \code{\link{add_totals_col}}
  \item \code{\link{add_totals_row}}
  \item \code{\link{remove_empty_rows}}
  \item \code{\link{remove_empty_cols}}
}
}
