% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empties.R
\name{remove_empty}
\alias{remove_empty}
\title{Remove empty rows and/or columns from a data.frame.}
\usage{
remove_empty(dat, which)
}
\arguments{
\item{dat}{the input data.frame.}

\item{which}{one of "row", "col", or \code{c("row", "col")}}
}
\value{
Returns the data.frame without its missing rows or columns.
}
\description{
Removes all rows and/or columns from a data.frame that are composed entirely of \code{NA} values.
}
\examples{
# not run:
# dat \%>\% remove_empty("rows")
}
