% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empties.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Removes empty columns from a data.frame.}
\usage{
remove_empty_cols(dat)
}
\arguments{
\item{dat}{the input data.frame.}
}
\value{
Returns the data.frame with no empty columns.
}
\description{
Removes all columns from a data.frame that are composed entirely of \code{NA} values.
}
\examples{
# called with magrittr pipe \%>\% :
# library(dplyr)
# not run:
# dat \%>\% remove_empty_cols

}

