% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janitor_deprecated.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Removes empty rows from a data.frame.}
\usage{
remove_empty_rows(dat)
}
\arguments{
\item{dat}{the input data.frame.}
}
\value{
Returns the data.frame with no empty rows.
}
\description{
This function is deprecated, use \code{remove_empty("rows")} instead.
}
\examples{
# not run:
# dat \%>\% remove_empty_rows
}
