% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_to_names.R
\name{row_to_names}
\alias{row_to_names}
\title{Elevate a row to be the column names of a data.frame}
\usage{
row_to_names(dat, row_number = 1, remove_row = TRUE,
  remove_rows_above = TRUE)
}
\arguments{
\item{dat}{The input data.frame}

\item{row_number}{The row of \code{dat} to make into the header}

\item{remove_row}{Remove \code{row_number} from the return value?}

\item{remove_rows_above}{If \code{row_number != 1}, remove rows
\code{1:(row_number-1)} from the return value?}
}
\value{
A data frame with new names (and some rows removed, if requested)
}
\description{
Elevate a row to be the column names of a data.frame
}
\examples{
data.frame(X__1=c(NA, "Title", 1:3),
           X__2=c(NA, "Title2", 4:6),
           stringsAsFactors=FALSE) \%>\%
  row_to_names(row_number=2)
}
