% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabyl.R
\name{tabyl}
\alias{tabyl}
\title{Generate a table of a vector.}
\usage{
tabyl(vec, show_na = TRUE, sort = FALSE)
}
\arguments{
\item{vec}{the vector to tabulate.}

\item{show_na}{should cases where the variable is NA be shown?}

\item{sort}{should the resulting table be sorted in descending order?}
}
\value{
Returns a data.frame (actually a \code{tbl_df}) with the frequencies of the tabulated variable.  Includes counts, percentages, and valid percentages (calculated omitting \code{NA} values, if present in the vector and \code{show_na = TRUE}.)
}
\description{
Get a frequency table of a variable as a data.frame, showing percentages and with or without including \code{NA} values.  A fully-featured alternative to \code{table()}.
}
\examples{
tabyl(mtcars$cyl)
tabyl(mtcars$cyl, sort = TRUE)
# called with magrittr pipe:
library(dplyr)
mtcars \%>\% .$cyl \%>\% tabyl()
# illustrating show_na functionality:
my_cars <- rbind(mtcars, rep(NA, 11))
tabyl(my_cars$cyl)
tabyl(my_cars$cyl, show_na = FALSE)
}

