% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_2.R
\name{top_2}
\alias{top_2}
\title{Generate a frequency table of a factor grouped by top-2, bottom-2, and all other levels.}
\usage{
top_2(input_vec, show_na = FALSE, sort = FALSE)
}
\arguments{
\item{input_vec}{the factor variable to tabulate.}

\item{show_na}{should cases where the variable is NA be shown?}

\item{sort}{should the resulting table be sorted in descending order?}
}
\value{
Returns a data.frame (actually a \code{tbl_df}) with the frequencies of the grouped, tabulated variable.  Includes counts and percentages, and valid percentages (calculated omitting \code{NA} values, if present in the vector and \code{show_na = TRUE}.)
}
\description{
Get a frequency table of a factor variable, grouped into level categories.
}
\examples{
top_2(as.factor(mtcars$hp))
}

